Working on semantics integration
At this point I'm not sure how to link the semantic objects
to geometry boundaries without duplicating the geometry. Duplicating
the geometry is a very bad idea because I would need to keep track
of the changes to the semantics/geometry at several locations, which
will fail for sure.

Ideally, I would use the same linking mechanism as it is in the data
model, thus the semantic objects would just point to parts of the
geometry boundary.

The other question is, how will it be possible that the user works
with a single citymodel, and cjio simply changes parts of that citymodel.
Instead of passing each cityobject into a distinct object and then
having to manually combine the citymodel. So somehow I need to keep
reference to every cityobject.