What if no SemanticSurface class?
Explore how could it work if there is no SemanticSurface class, but
the semantic information is stored in the Geometry class and the
semantic surfaces are extracted on demand by a Geometry.get_surfaces()
method.

However, this alternative does not allow to get the children and parents
of the individual semantic surface objects.