Add bumpversion config
+ Adds a bumpversion configuration for automatically updating
the version numbers in the setup script, package and
documentation.
+ Requires bump2version (the maintained fork of bumpversion).
+ Replaces the dynamic version number in the setup script with
a static string. This makes it possible to retrieve the package
version without importing the package, and thus easier to
integrate with automation tools. Since bump2version takes care
of updating the version numbers in all locations, its not a
problem that they are specified in multiple locations, provided
that bump2version is used for each version bump.
+ Creates a dedicated commit and a tag for each version. Can be
changed in setup.cfg:bumpversion.