Implement new behaviour for +metadata-extended
The new behaviour is that is +metadata-extended is in the file then
ops that modify the file (merge, subset, filter) will update the
+metadata-extended, if not present then all ignored.

No granularity, it's all of nothing (including lineage).

One can simply add the +metadata-extended with the new cjio operator:
metadata_create

And one can delete the +metadata-extended (which would mean the ops
won't store or update metadata) with the new cjio operator:
metadata_remove

I also bluntly decided to start the name of cjio with the property and
add the verb *after*, easier for users, so we have for metadata:
```
metadata_create            Add the +metadata-extended properties...
metadata_get               Shows the metadata and +metadata-extended of...
metadata_remove            Remove the +metadata-extended properties.
metadata_update            Update the +metadata-extended for properties...
```