Renamed a lot of functions to make things easier for the users.
It was a mess and difficult to find, but now it's [value+action] where
possible.

I also remove some like remove_duplicate_vertices() because vertices_clean()
is calling it. If someone wants a finer granularity she can use the API.

The names of the functions in the API are the same for now to avoid
breaking the whole thing. I guess we should change them too?

I also renamed all "epsg" to "crs": clearer.