Remove the flat that enforced all CRS to have xy in that order
this gave issues for CRS where lat-long is used, and the further conversion
was crashing with 'inf'...

If I take a dutch building in 7415, then convert to 4937 I get lat-long.

Then again convert back to 7415 and all is good.

So I *think* it works, but more tests are surely needed :J